Validator class moved to shared module from control
